update: move Earth to planet
