+++ planetary-orbital-calculator/src/test/java/com/tdd/PlanetOrbitTest.java
@@ -1,5 +1,6 @@
 package com.tdd;
 
+import com.tdd.planet.EARTH;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
